src/data_generator/requirements.txt
boto3
paho-mqtt